New module kas-core::event::cx::send
